Add informal diagram consistency pipeline steps (#308)
* Adapt code model to support nested classes

* Extend code model to store references of types to other types

* Move code from thesis repo into ArDoCo core

* Fix some test

* Merge the two existing diagram classes

* Fix the large graph test

* Fix version name

* Update stages/diagram-consistency/pom.xml


* Adapt pom.xml files as reviewed and adapt to changes related to package structure

* Update framework/common/pom.xml

* Apply spotless

* Reset benchmark

* Squashed 'tests/tests-base/src/main/resources/benchmark/' changes from 7675243a..daeaee7e

daeaee7e Update code models to contain parent class and reference info (#12)
4fae195e Format Code Model. Fix #13
814aa945 Remove duplicate in GS

git-subtree-dir: tests/tests-base/src/main/resources/benchmark
git-subtree-split: daeaee7e6273722b44a1a82b5d7bdd204edd743c

* Let evaluation code extract code models if no already extracted model available

* Apply spotless

* Use assertion messages to gain more information about the occurring failures

* When extracting a code model, clone first

* Delete directory before cloning

* Squashed 'tests/tests-base/src/main/resources/benchmark/' changes from daeaee7e..9bbc3a89

9bbc3a89 Update code models to contain parent class and reference info, using new formatting (#16)
fe5a8549 Update the codemodel to be more (human) readable

git-subtree-dir: tests/tests-base/src/main/resources/benchmark
git-subtree-split: 9bbc3a89677ffba98bb861f0c0206a93c61ec5d1

* Work on code smells

* Remove annotation that leads to compile error

* Register diagram-consistency to report

* Fix loading of ACM

* Remove commented out import

---------

+++ tests/tests-base/src/main/java/edu/kit/kastel/mcse/ardoco/core/tests/eval/EvaluationMetrics.java
@@ -203,7 +203,7 @@ public class EvaluationMetrics {
      * @return the specificity
      */
     public static double calculateSpecificity(int trueNegatives, int falsePositives) {
-        double specificity = ((double) trueNegatives) / ((double) trueNegatives + falsePositives);
+        double specificity = trueNegatives / ((double) trueNegatives + falsePositives);
         if (Double.isNaN(specificity)) {
             return 1.0;
         }
